✨ feat(screens): add main screen and navigation system
- The main screen was added, responsible for displaying the frames: Main, Consult and Profile.
- The dependencies of the Bottom-tabs were used to navigate between the frames.

Bugs with the version: In the navigation tabs container, it is not possible to add the "Main background color" behind the container, causing a break in color continuity with the application background.

Temporary solution: The value of the container's "borderRound" property was reduced, reducing the perception of a break in continuity.