✨feat(domain): added Email and CelPhone value object classes
The Email class receives a string in the "create" method, returning an entity of the email type. The constructor is private and validation is done in the create method process, ensuring that when the value is obtained by the "getValue" method, it is ready for correct use.

The CelPhone class receives a string in the "create" method, returning an entity of the CelPhone type. The constructor is private and validation and formatting is done in the create method process, ensuring that when the value is obtained by the "getValue" method, it is ready for correct use. In this class, there are two ways to obtain the value, using the "getValue" method, which returns a string containing only numeric digits in the Brazilian standard, containing the country's number (55). This method is intended to provide information in a format aimed at persistence. The "getFormatedValue" method returns a formatted string in a more humanized way, containing special characters and spaces. However, it still follows the Brazilian standard, being intended for use in API responses.