✨feat(domain): added Role and DateTools value object script
The Role script provides functions that help in validating and obtaining values ​​and keys by switching between them. It also exposes the Role enum.

The DateTools script provides functions that help in converting the date class to the project's date and time format and vice versa. The information is separated and in the date (yyyy-MM-dd) and time (HH:MM) formats. Validation tools are also provided.