✨feat(interfaces): added route class
A route file has been added using the Express Router class. All routes are described in this file in order to centralize responsibility. Using AuthMiddleware in the route, this becomes a private route, capturing the identifier from the token before executing the controller.

Attention: so far in the project, only user resources have private routes.