✨feat(component): Updated Doctor List item component
The ability to execute a function passed as a reference via parameter has been implemented. The responsibility for accessing the icon has been passed to this component, receiving a character F or M via parameter, being of type "string". The doctor's ID has been passed as a parameter and used as a parameter of the function passed in "onPress".

The purpose of this component is to render a component that displays the doctor's information and executes an action when pressed.