✨feat(component): Added a functional component for a general confirmation modal
This modal receives two "strings", one for the title and one for the body, and also two callback functions for confirming and one for canceling. In addition to controlling visibility via parameter, receiving a boolean value.

The goal is to render a component that can be reused for different purposes that require user interaction requiring validation by the user. In short, it is a component used to consult the user's opinion with yes or no questions (Example: Do you really want to delete your account? With the options to confirm or cancel)