✨feat(component): Updated the Appointment List item component
The ability to execute a function passed as a reference via parameter has been implemented. The date string has been formatted to the Brazilian standard (dd/mm/yyyy). The Appointment ID has been passed as a parameter and used as a parameter of the function passed in "onPress".

The purpose of this component is to render a component that displays the doctor's information and executes an action when the "Cancel" button is pressed.