✨feat(component): Updated the Service List item component
The ability to execute a function passed as a reference via parameter has been implemented.

The component has also received new parameters that are used as action parameters and the description as the body for a general modal.

The purpose of this component is to render a component that displays the service information and perform a navigation action to the general modal screen when the "?" button is pressed and to the "Scheduling" screen when "Book" is pressed.