✨feat(screen): Updated Home screen
The ability to retrieve data from the server via request has been implemented. If the process of loading doctors is completed successfully, the status is updated from "undefined" to the list of doctors. The status of the doctors is used to define whether a list of doctors or an informative message will be displayed. The process is repeated every time the component is assembled.

Changes were made to the value passing to the "icon" property due to the transfer of responsibility to the component that renders the item itself.

Each item component will receive a "SelectDoctor" function, which will have the parameters set by the values ​​of each component. The function consists of navigation and transferring the specific parameters to the Services screen.

The style file of the screen component received the addition of a style for the display message when an error occurs when obtaining doctors.