✨feat(screen): Updated Profile frame
The ability to obtain data from the server via request has been implemented. If the user loading process is completed successfully, the status is updated from "undefined" to the userProfile type (defined momentarily in the file itself). The userProfile status is used to define whether the information or an informative message will be displayed. The process is repeated every time the component is assembled.

The "Sair" button has been added to perform the logout process by obtaining the function from the authentication context and passing it as the value of the "onPress" property.

The screen component's style file has received the addition of a style for the message displayed when an error occurs when obtaining doctors.