✨feat(screen): Updated Services screen
The ability to obtain data from the server via request has been implemented. If the service loading process is completed successfully, the status is updated from "undefined" to the list of services of the 'IServicesList' type (type described in the file itself). The status of the services is used to define whether a list of services or an informative message will be displayed. The process is repeated every time the component is assembled.
The icon's layout has been changed since the path is no longer passed as a parameter and the responsibility for loading the information lies with this component.
Each item component will receive a "toSchedule" function, which will have the parameters set by the values ​​of each component. The function consists of browsing and transferring the specific parameters (id_doctor, id_service) to the Scheduling screen.

The screen component's style file has received the addition of a style for the display message when an error occurs when obtaining doctors.